+++ src/network/src/downloader/piece_info.rs
@@ -22,7 +22,7 @@ use chain::{MAX_TX_SET_SIZE, MAX_PIECE_SIZE, MAX_SUB_PIECE_SIZE};
 #[derive(Debug, Clone, PartialEq)]
 pub struct PieceInfo {
     /// The size of the piece
-    pub(crate) size: usize,
+    pub(crate) size: u64,
 
     /// Sub-pieces info
     pub(crate) sub_pieces: Vec<SubPieceInfo>,
@@ -37,14 +37,14 @@ impl PieceInfo {
         let mut size = 0;
 
         for info in sub_pieces.iter() {
-            if info.size > MAX_SUB_PIECE_SIZE {
+            if info.size > MAX_SUB_PIECE_SIZE as u64 {
                 panic!("Cannot crate PieceInfo out of a SubPiece size greater than {}! Got: {}", MAX_SUB_PIECE_SIZE, info.size);
             }
 
             size += info.size;
         }
 
-        if size > MAX_PIECE_SIZE {
+        if size > MAX_PIECE_SIZE as u64 {
             panic!("Cannot crate PieceInfo with a sum of all SubPieces sizes greater than {}! Got: {}", MAX_PIECE_SIZE, size);
         }
